django_simple_requirements.txt
Django==4.2.7
Pillow==9.5.0
whitenoise==6.6.0
torch==2.0.1
torchvision==0.15.2
timm==0.9.7
numpy==1.24.3
opencv-python==4.8.0.76
scikit-learn==1.3.0
matplotlib==3.7.2